requirements.txt
pyaudio~=0.2.14
openai~=1.3.3
pygame~=2.5.2
python-dotenv~=1.0.0